support whitelisted slash commands in the agents view
The agents view previously sent all input verbatim, so typing /login
created a junk agent named "/login". Now /login, /logout, /model, and
/quit run directly in the view, other built-in commands surface a hint
to open an agent first, and unknown /text still passes through for
daemon-side template/skill expansion.

Provider auth flows (login/logout dialogs and selectors) move out of
InteractiveMode into a reusable ProviderAuthFlows class so both views
share the same UI; /logout in interactive mode now renders as a
full-pane overlay like /login. The agents view editor also gains slash
command autocomplete, and /model picks the default model for new
agents.
