+++ packages/coding-agent/src/modes/interactive/auth-flows.ts
@@ -1,0 +1,682 @@
+/**
+ * Provider authentication flows (login/logout dialogs and selectors).
+ *
+ * Extracted from InteractiveMode so any TUI surface that owns a `TUI` and a
+ * `ModelRegistry` (interactive sessions, the agents view) can run the same
+ * auth UI. Host-specific side effects (footer refresh, billing warnings) hang
+ * off the small host interface instead of the flows themselves.
+ */
+
+import * as path from "node:path";
+import { getProviders, type OAuthProviderId, type OAuthSelectPrompt } from "@earendil-works/pi-ai";
+import type { OverlayHandle, TUI } from "@earendil-works/pi-tui";
+import { getAuthPath, getDocsPath } from "../../config.js";
+import type { ModelRegistry } from "../../core/model-registry.js";
+import {
+	checkPrimeInferenceAccess,
+	fetchPrimeTeams,
+	loadPrimeCliConfig,
+	loginPrimeInference,
+	PRIME_INFERENCE_PROVIDER_ID,
+	PRIME_INFERENCE_PROVIDER_NAME,
+	type PrimeTeam,
+} from "../../core/prime-inference-auth.js";
+import { BUILT_IN_PROVIDER_DISPLAY_NAMES } from "../../core/provider-display-names.js";
+import { showFullPaneOverlay } from "./components/centered-overlay.js";
+import { ExtensionSelectorComponent } from "./components/extension-selector.js";
+import { LoginDialogComponent } from "./components/login-dialog.js";
+import {
+	type AuthSelectorProvider,
+	compareAuthSelectorProviders,
+	OAuthSelectorComponent,
+} from "./components/oauth-selector.js";
+import { PrimeTeamSelectorComponent } from "./components/prime-team-selector.js";
+import { theme } from "./theme/theme.js";
+
+export type AuthenticationResult =
+	| {
+			status: "success";
+			providerId: string;
+			providerName: string;
+			authType: "oauth" | "api_key";
+	  }
+	| { status: "cancelled" }
+	| { status: "failed" };
+
+export const BEDROCK_PROVIDER_ID = "amazon-bedrock";
+
+const BUILT_IN_MODEL_PROVIDERS = new Set<string>(getProviders());
+
+export function isApiKeyLoginProvider(
+	providerId: string,
+	oauthProviderIds: ReadonlySet<string>,
+	builtInProviderIds: ReadonlySet<string> = BUILT_IN_MODEL_PROVIDERS,
+): boolean {
+	if (BUILT_IN_PROVIDER_DISPLAY_NAMES[providerId]) {
+		return true;
+	}
+	if (builtInProviderIds.has(providerId)) {
+		return false;
+	}
+	return !oauthProviderIds.has(providerId);
+}
+
+export interface ProviderAuthFlowsHost {
+	readonly ui: TUI;
+	readonly modelRegistry: ModelRegistry;
+	showStatus(message: string): void;
+	showError(message: string): void;
+	/** Models currently visible to the host; used to detect providers configured via external credentials. */
+	getAvailableModels(): Promise<ReadonlyArray<{ provider: string }>>;
+	/** Invoked after stored credentials change so the host can refresh dependent UI. */
+	onAuthChanged?(): void | Promise<void>;
+	/** Invoked after a successful login (e.g. to surface billing warnings). */
+	onLoginCompleted?(): void;
+}
+
+export class ProviderAuthFlows {
+	constructor(private readonly host: ProviderAuthFlowsHost) {}
+
+	/** Runs the provider selector followed by the matching login dialog. */
+	runLogin(authType?: "oauth" | "api_key"): Promise<AuthenticationResult> {
+		const providerOptions = this.getLoginProviderOptions(authType);
+		if (providerOptions.length === 0) {
+			this.host.showStatus(
+				authType === "oauth"
+					? "No subscription providers available."
+					: authType === "api_key"
+						? "No API key providers available."
+						: "No providers available.",
+			);
+			return Promise.resolve({ status: "failed" });
+		}
+
+		return new Promise((resolve) => {
+			let handle: OverlayHandle | undefined;
+			const close = () => {
+				handle?.hide();
+				this.host.ui.requestRender();
+			};
+			const selector = new OAuthSelectorComponent(
+				"login",
+				this.host.modelRegistry.authStorage,
+				providerOptions,
+				async (providerOption: AuthSelectorProvider) => {
+					close();
+
+					if (providerOption.authType === "oauth") {
+						resolve(await this.showLoginDialog(providerOption.id, providerOption.name));
+					} else if (providerOption.id === PRIME_INFERENCE_PROVIDER_ID) {
+						resolve(await this.runPrimeInferenceLogin());
+					} else if (providerOption.id === BEDROCK_PROVIDER_ID) {
+						resolve(await this.showBedrockSetupDialog(providerOption.id, providerOption.name));
+					} else {
+						resolve(await this.showApiKeyLoginDialog(providerOption.id, providerOption.name));
+					}
+				},
+				() => {
+					close();
+					resolve({ status: "cancelled" });
+				},
+				(providerId) => this.host.modelRegistry.getProviderAuthStatus(providerId),
+				{ getRows: () => this.host.ui.terminal.rows },
+			);
+			handle = showFullPaneOverlay(this.host.ui, selector, 78);
+		});
+	}
+
+	/** Shows the stored-credential selector and removes the chosen credential. */
+	runLogout(): Promise<void> {
+		const providerOptions = this.getLogoutProviderOptions();
+		if (providerOptions.length === 0) {
+			this.host.showStatus(
+				"No stored credentials to remove. /logout only removes credentials saved by /login; environment variables and models.json config are unchanged.",
+			);
+			return Promise.resolve();
+		}
+
+		return new Promise((resolve) => {
+			let handle: OverlayHandle | undefined;
+			const close = () => {
+				handle?.hide();
+				this.host.ui.requestRender();
+			};
+			const selector = new OAuthSelectorComponent(
+				"logout",
+				this.host.modelRegistry.authStorage,
+				providerOptions,
+				async (providerOption: AuthSelectorProvider) => {
+					close();
+
+					try {
+						this.host.modelRegistry.authStorage.logout(providerOption.id);
+						this.host.modelRegistry.refresh();
+						await this.host.onAuthChanged?.();
+						const message =
+							providerOption.authType === "oauth"
+								? `Logged out of ${providerOption.name}`
+								: `Removed stored API key for ${providerOption.name}. Environment variables and models.json config are unchanged.`;
+						this.host.showStatus(message);
+					} catch (error: unknown) {
+						this.host.showError(`Logout failed: ${error instanceof Error ? error.message : String(error)}`);
+					}
+					resolve();
+				},
+				() => {
+					close();
+					resolve();
+				},
+				undefined,
+				{ getRows: () => this.host.ui.terminal.rows },
+			);
+			handle = showFullPaneOverlay(this.host.ui, selector, 78);
+		});
+	}
+
+	private getLoginProviderOptions(authType?: "oauth" | "api_key"): AuthSelectorProvider[] {
+		const authStorage = this.host.modelRegistry.authStorage;
+		const oauthProviders = authStorage.getOAuthProviders();
+		const oauthProviderIds = new Set(oauthProviders.map((provider) => provider.id));
+		const options: AuthSelectorProvider[] = oauthProviders.map((provider) => ({
+			id: provider.id,
+			name: provider.name,
+			authType: "oauth",
+		}));
+
+		const modelProviders = new Set(this.host.modelRegistry.getAll().map((model) => model.provider));
+		for (const providerId of modelProviders) {
+			if (!isApiKeyLoginProvider(providerId, oauthProviderIds)) {
+				continue;
+			}
+			options.push({
+				id: providerId,
+				name: this.host.modelRegistry.getProviderDisplayName(providerId),
+				authType: "api_key",
+			});
+		}
+
+		const filteredOptions = authType ? options.filter((option) => option.authType === authType) : options;
+		return filteredOptions.sort(compareAuthSelectorProviders);
+	}
+
+	private getLogoutProviderOptions(): AuthSelectorProvider[] {
+		const authStorage = this.host.modelRegistry.authStorage;
+		const options: AuthSelectorProvider[] = [];
+
+		for (const providerId of authStorage.list()) {
+			const credential = authStorage.get(providerId);
+			if (!credential) {
+				continue;
+			}
+			options.push({
+				id: providerId,
+				name: this.host.modelRegistry.getProviderDisplayName(providerId),
+				authType: credential.type,
+			});
+		}
+
+		return options.sort((a, b) => a.name.localeCompare(b.name));
+	}
+
+	private async completeProviderAuthentication(
+		providerId: string,
+		providerName: string,
+		authType: "oauth" | "api_key",
+		statusSuffix?: string,
+	): Promise<AuthenticationResult> {
+		this.host.modelRegistry.refresh();
+
+		const actionLabel = authType === "oauth" ? `Logged in to ${providerName}` : `Saved API key for ${providerName}`;
+		await this.host.onAuthChanged?.();
+		this.host.showStatus(
+			`${actionLabel}. Credentials saved to ${getAuthPath()}${statusSuffix ? `. ${statusSuffix}` : ""}`,
+		);
+		this.host.onLoginCompleted?.();
+		return {
+			status: "success",
+			providerId,
+			providerName,
+			authType,
+		};
+	}
+
+	private async completeExternalProviderSetup(
+		providerId: string,
+		providerName: string,
+	): Promise<AuthenticationResult> {
+		this.host.modelRegistry.refresh();
+		await this.host.onAuthChanged?.();
+		this.host.showStatus(`${providerName} uses external credentials. Select a model after configuring them.`);
+		return {
+			status: "success",
+			providerId,
+			providerName,
+			authType: "api_key",
+		};
+	}
+
+	private async hasAvailableProviderModels(providerId: string): Promise<boolean> {
+		const models = await this.host.getAvailableModels();
+		return models.some((model) => model.provider === providerId);
+	}
+
+	private async showBedrockSetupDialog(providerId: string, providerName: string): Promise<AuthenticationResult> {
+		const dialog = new LoginDialogComponent(
+			this.host.ui,
+			providerId,
+			() => {
+				// Completion handled below.
+			},
+			providerName,
+			"Amazon Bedrock setup",
+		);
+		const handle = showFullPaneOverlay(this.host.ui, dialog, 88);
+		const closeDialog = () => {
+			handle.hide();
+			this.host.ui.requestRender();
+		};
+
+		try {
+			await dialog.showContinueInfo([
+				theme.fg("text", "Amazon Bedrock uses AWS credentials instead of a single API key."),
+				theme.fg("text", "Configure an AWS profile, IAM keys, bearer token, or role-based credentials."),
+				theme.fg("muted", "See:"),
+				theme.fg("accent", `  ${path.join(getDocsPath(), "providers.md")}`),
+			]);
+			closeDialog();
+			if (!(await this.hasAvailableProviderModels(providerId))) {
+				this.host.showStatus(`${providerName} credentials were not detected. Configure them, then reopen /model.`);
+				return { status: "cancelled" };
+			}
+			return await this.completeExternalProviderSetup(providerId, providerName);
+		} catch (error: unknown) {
+			closeDialog();
+			const errorMsg = error instanceof Error ? error.message : String(error);
+			if (errorMsg !== "Login cancelled") {
+				this.host.showError(`Failed to set up ${providerName}: ${errorMsg}`);
+				return { status: "failed" };
+			}
+			return { status: "cancelled" };
+		}
+	}
+
+	private showPrimeTeamSelector(
+		teams: PrimeTeam[],
+		currentTeamId: string | undefined,
+	): Promise<PrimeTeam | null | undefined> {
+		return new Promise((resolve) => {
+			let handle: OverlayHandle | undefined;
+			const close = () => {
+				handle?.hide();
+				this.host.ui.requestRender();
+			};
+			const selector = new PrimeTeamSelectorComponent(
+				teams,
+				currentTeamId,
+				(team) => {
+					close();
+					resolve(team);
+				},
+				() => {
+					close();
+					resolve(undefined);
+				},
+				{ getRows: () => this.host.ui.terminal.rows },
+			);
+			handle = showFullPaneOverlay(this.host.ui, selector, 78);
+		});
+	}
+
+	private getPrimeInferenceDefaultTeamStatus(): string {
+		const storedTeam = this.host.modelRegistry.authStorage.getPrimeInferenceTeamSelection();
+		if (storedTeam) {
+			return `Using team "${storedTeam.name}".`;
+		}
+		if (storedTeam === null) {
+			return "Using personal account.";
+		}
+		let config: ReturnType<typeof loadPrimeCliConfig>;
+		try {
+			config = loadPrimeCliConfig();
+		} catch {
+			return "Using personal account.";
+		}
+		if (config.teamIdFromEnv) {
+			return "Using team from PRIME_TEAM_ID.";
+		}
+		if (config.teamName) {
+			return `Using team "${config.teamName}".`;
+		}
+		if (config.teamId) {
+			return "Using Prime CLI team.";
+		}
+		return "Using personal account.";
+	}
+
+	private async selectPrimeInferenceTeam(apiKey: string, dialog: LoginDialogComponent): Promise<string | undefined> {
+		try {
+			const config = loadPrimeCliConfig();
+			if (config.teamIdFromEnv) {
+				this.host.modelRegistry.authStorage.reload();
+				return "Using team from PRIME_TEAM_ID.";
+			}
+
+			dialog.showProgress("Loading Prime teams...");
+			const teams = await fetchPrimeTeams(apiKey, config.baseUrl, { signal: dialog.signal });
+			if (dialog.signal.aborted) {
+				return this.getPrimeInferenceDefaultTeamStatus();
+			}
+			if (teams.length === 0) {
+				this.host.modelRegistry.authStorage.setPrimeInferenceTeamSelection(null);
+				return "Using personal account.";
+			}
+
+			const storedTeam = this.host.modelRegistry.authStorage.getPrimeInferenceTeamSelection();
+			const currentTeamId = storedTeam === null ? undefined : (storedTeam?.teamId ?? config.teamId);
+			const selectedTeam = await this.showPrimeTeamSelector(teams, currentTeamId);
+			if (selectedTeam !== undefined) {
+				this.host.modelRegistry.authStorage.setPrimeInferenceTeamSelection(selectedTeam);
+			}
+			return selectedTeam
+				? `Using team "${selectedTeam.name}".`
+				: selectedTeam === null
+					? "Using personal account."
+					: this.getPrimeInferenceDefaultTeamStatus();
+		} catch {
+			this.host.modelRegistry.authStorage.reload();
+			return this.getPrimeInferenceDefaultTeamStatus();
+		}
+	}
+
+	private async completePrimeInferenceLogin(
+		apiKey: string,
+		dialog: LoginDialogComponent,
+		closeDialog: () => void,
+	): Promise<AuthenticationResult> {
+		const previousPrimeCredential = this.host.modelRegistry.authStorage.get(PRIME_INFERENCE_PROVIDER_ID);
+		const previousPrimeTeam =
+			previousPrimeCredential?.type === "api_key" ? previousPrimeCredential.primeTeam : undefined;
+		this.host.modelRegistry.authStorage.set(PRIME_INFERENCE_PROVIDER_ID, {
+			type: "api_key",
+			key: apiKey,
+			...(previousPrimeTeam !== undefined ? { primeTeam: previousPrimeTeam } : {}),
+		});
+		const teamStatus = await this.selectPrimeInferenceTeam(apiKey, dialog);
+
+		closeDialog();
+		return await this.completeProviderAuthentication(
+			PRIME_INFERENCE_PROVIDER_ID,
+			PRIME_INFERENCE_PROVIDER_NAME,
+			"api_key",
+			teamStatus,
+		);
+	}
+
+	async runPrimeInferenceLogin(): Promise<AuthenticationResult> {
+		const dialog = new LoginDialogComponent(
+			this.host.ui,
+			PRIME_INFERENCE_PROVIDER_ID,
+			(_success, _message) => {
+				// Completion handled below.
+			},
+			PRIME_INFERENCE_PROVIDER_NAME,
+		);
+
+		const handle = showFullPaneOverlay(this.host.ui, dialog, 88);
+
+		const closeDialog = () => {
+			handle.hide();
+			this.host.ui.requestRender();
+		};
+
+		// The browser challenge gets its own controller so a manually pasted key
+		// can stop the polling without tearing down the dialog.
+		const browserAbort = new AbortController();
+		const onDialogAbort = () => browserAbort.abort();
+		dialog.signal.addEventListener("abort", onDialogAbort, { once: true });
+
+		let manualInputArmed = false;
+		let resolveManualKey: (entry: { apiKey: string; source: "manual" }) => void = () => {};
+		const manualKeyEntry = new Promise<{ apiKey: string; source: "manual" }>((resolve) => {
+			resolveManualKey = resolve;
+		});
+		const armManualInput = (prompt: string): void => {
+			manualInputArmed = true;
+			void (async () => {
+				let value = (await dialog.showManualInput(prompt)).trim();
+				while (!value) {
+					value = (await dialog.waitForInput()).trim();
+				}
+				resolveManualKey({ apiKey: value, source: "manual" });
+			})().catch(() => {
+				// Cancellation surfaces through the dialog signal.
+			});
+		};
+
+		try {
+			const browserLogin = loginPrimeInference({
+				onAuth: (info) => {
+					dialog.showAuth(info.url, info.instructions);
+					armManualInput("Complete the sign-in in your browser, or paste an API key below:");
+				},
+				onProgress: (message) => {
+					dialog.showProgress(message);
+				},
+				signal: browserAbort.signal,
+			});
+			// When the browser challenge cannot start or breaks down, keep the dialog
+			// open and fall back to plain API key entry instead of failing outright.
+			const browserLoginOrFallback = browserLogin.catch((error: unknown) => {
+				if (browserAbort.signal.aborted) {
+					throw error;
+				}
+				const errorMsg = error instanceof Error ? error.message : String(error);
+				dialog.showProgress(`Browser sign-in unavailable (${errorMsg}).`);
+				if (!manualInputArmed) {
+					armManualInput("Paste a Prime API key below:");
+				}
+				return manualKeyEntry;
+			});
+			// Once the browser flow has settled into manual fallback, nothing above
+			// rejects on cancel anymore, so the dialog signal must end the race too.
+			const dialogCancelled = new Promise<never>((_, reject) => {
+				dialog.signal.addEventListener("abort", () => reject(new Error("Login cancelled")), { once: true });
+			});
+			// Promise.race observes the rejections below, but keep dedicated handlers
+			// so neither an aborted browser flow nor a cancelled dialog can surface
+			// as an unhandled rejection.
+			browserLoginOrFallback.catch(() => {});
+			dialogCancelled.catch(() => {});
+
+			const result = await Promise.race([browserLoginOrFallback, manualKeyEntry, dialogCancelled]);
+			if (dialog.signal.aborted) {
+				closeDialog();
+				return { status: "cancelled" };
+			}
+
+			if (result.source === "manual") {
+				browserAbort.abort();
+				dialog.showProgress("Checking Prime Inference access...");
+				const config = loadPrimeCliConfig();
+				const access = await checkPrimeInferenceAccess(result.apiKey, config.baseUrl, { signal: dialog.signal });
+				if (dialog.signal.aborted) {
+					closeDialog();
+					return { status: "cancelled" };
+				}
+				if (!access.ok) {
+					const status = access.status === undefined ? "" : `HTTP ${access.status}: `;
+					throw new Error(`Prime API key does not have Prime Inference access (${status}${access.message})`);
+				}
+			}
+
+			return await this.completePrimeInferenceLogin(result.apiKey, dialog, closeDialog);
+		} catch (error: unknown) {
+			closeDialog();
+			const errorMsg = error instanceof Error ? error.message : String(error);
+			if (!dialog.signal.aborted && errorMsg !== "Login cancelled") {
+				this.host.showError(`Failed to login to ${PRIME_INFERENCE_PROVIDER_NAME}: ${errorMsg}`);
+				return { status: "failed" };
+			}
+			return { status: "cancelled" };
+		} finally {
+			dialog.signal.removeEventListener("abort", onDialogAbort);
+		}
+	}
+
+	private async showApiKeyLoginDialog(providerId: string, providerName: string): Promise<AuthenticationResult> {
+		const dialog = new LoginDialogComponent(
+			this.host.ui,
+			providerId,
+			(_success, _message) => {
+				// Completion handled below
+			},
+			providerName,
+		);
+
+		const handle = showFullPaneOverlay(this.host.ui, dialog, 88);
+
+		const closeDialog = () => {
+			handle.hide();
+			this.host.ui.requestRender();
+		};
+
+		try {
+			const apiKey = (await dialog.showPrompt("Enter API key:")).trim();
+			if (!apiKey) {
+				throw new Error("API key cannot be empty.");
+			}
+
+			this.host.modelRegistry.authStorage.set(providerId, { type: "api_key", key: apiKey });
+
+			closeDialog();
+			return await this.completeProviderAuthentication(providerId, providerName, "api_key");
+		} catch (error: unknown) {
+			closeDialog();
+			const errorMsg = error instanceof Error ? error.message : String(error);
+			if (errorMsg !== "Login cancelled") {
+				this.host.showError(`Failed to save API key for ${providerName}: ${errorMsg}`);
+				return { status: "failed" };
+			}
+			return { status: "cancelled" };
+		}
+	}
+
+	private showOAuthLoginSelect(dialogHandle: OverlayHandle, prompt: OAuthSelectPrompt): Promise<string | undefined> {
+		return new Promise((resolve) => {
+			dialogHandle.setHidden(true);
+			let selectorHandle: OverlayHandle | undefined;
+			const restoreDialog = () => {
+				selectorHandle?.hide();
+				dialogHandle.setHidden(false);
+				dialogHandle.focus();
+				this.host.ui.requestRender();
+			};
+			const labels = prompt.options.map((option) => option.label);
+			const selector = new ExtensionSelectorComponent(
+				prompt.message,
+				labels,
+				(optionLabel) => {
+					restoreDialog();
+					resolve(prompt.options.find((option) => option.label === optionLabel)?.id);
+				},
+				() => {
+					restoreDialog();
+					resolve(undefined);
+				},
+				{ getRows: () => this.host.ui.terminal.rows },
+			);
+			selectorHandle = showFullPaneOverlay(this.host.ui, selector, 76);
+		});
+	}
+
+	private async showLoginDialog(providerId: string, providerName: string): Promise<AuthenticationResult> {
+		const providerInfo = this.host.modelRegistry.authStorage
+			.getOAuthProviders()
+			.find((provider) => provider.id === providerId);
+
+		// Providers that use callback servers (can paste redirect URL)
+		const usesCallbackServer = providerInfo?.usesCallbackServer ?? false;
+
+		// Create login dialog component
+		const dialog = new LoginDialogComponent(
+			this.host.ui,
+			providerId,
+			(_success, _message) => {
+				// Completion handled below
+			},
+			providerName,
+		);
+
+		const dialogHandle = showFullPaneOverlay(this.host.ui, dialog, 88);
+
+		// Promise for manual code input (racing with callback server)
+		let manualCodeResolve: ((code: string) => void) | undefined;
+		let manualCodeReject: ((err: Error) => void) | undefined;
+		const manualCodePromise = new Promise<string>((resolve, reject) => {
+			manualCodeResolve = resolve;
+			manualCodeReject = reject;
+		});
+
+		// Close dialog overlay helper.
+		const closeDialog = () => {
+			dialogHandle.hide();
+			this.host.ui.requestRender();
+		};
+
+		try {
+			await this.host.modelRegistry.authStorage.login(providerId as OAuthProviderId, {
+				onAuth: (info: { url: string; instructions?: string }) => {
+					dialog.showAuth(info.url, info.instructions);
+
+					if (usesCallbackServer) {
+						// Show input for manual paste, racing with callback
+						dialog
+							.showManualInput("Paste redirect URL below, or complete login in browser:")
+							.then((value) => {
+								if (value && manualCodeResolve) {
+									manualCodeResolve(value);
+									manualCodeResolve = undefined;
+								}
+							})
+							.catch(() => {
+								if (manualCodeReject) {
+									manualCodeReject(new Error("Login cancelled"));
+									manualCodeReject = undefined;
+								}
+							});
+					} else if (providerId === "github-copilot") {
+						// GitHub Copilot polls after onAuth
+						dialog.showWaiting("Waiting for browser authentication...");
+					}
+					// For Anthropic: onPrompt is called immediately after
+				},
+
+				onPrompt: async (prompt: { message: string; placeholder?: string }) => {
+					return dialog.showPrompt(prompt.message, prompt.placeholder);
+				},
+
+				onProgress: (message: string) => {
+					dialog.showProgress(message);
+				},
+
+				onSelect: (prompt: OAuthSelectPrompt) => this.showOAuthLoginSelect(dialogHandle, prompt),
+
+				onManualCodeInput: () => manualCodePromise,
+
+				signal: dialog.signal,
+			});
+
+			// Success
+			closeDialog();
+			return await this.completeProviderAuthentication(providerId, providerName, "oauth");
+		} catch (error: unknown) {
+			closeDialog();
+			const errorMsg = error instanceof Error ? error.message : String(error);
+			if (errorMsg !== "Login cancelled") {
+				this.host.showError(`Failed to login to ${providerName}: ${errorMsg}`);
+				return { status: "failed" };
+			}
+			return { status: "cancelled" };
+		}
+	}
+}
